Revert "[Frontend] Factor out code for running uvicorn (#6828)"
This reverts commit 981b0d567355063d5453e382a85970cae083c615.
