--- vllm/server/__init__.py
@@ -1,3 +1,0 @@
-from .launch import serve_http
-
-__all__ = ["serve_http"]
